✨ final review (#169)
* ♻️ refactor: separate ThemeContext and useTheme hook (#160)

* ♻️ refactor: separate ThemeContext and useTheme hook (#160)

* 🔧 fix: remove console.error in NotFoundPage (#160)

* ✨ feat: add @Slf4j and @Transactional annotations to services (#160)

* 🔧 fix: update imports and barrel exports (#160)

* ♻️ refactor: update FavoriteDto to follow DTO pattern (#160)

* 📝 docs: sync README with actual implementation (#160)

* 💄 design: add favicon (#160)

* ✨ feat: hide EcosystemPlaceholder (#160)
