+++ README.md
@@ -36,11 +36,13 @@ Crypto Market Insight는 가상자산 시장 데이터를 기반으로
 - 타임프레임별 가격 추세 분석
 - 거래량 및 변동성 지표 제공
 
-### 3.2 체인 / 생태계 기반 분류
+### 3.2 체인 / 생태계 기반 분류 (Coming Soon)
 
 - BTC, ETH(ERC-20), SOL 등 주요 체인 기준 코인 분류
 - 생태계별 시장 흐름 요약 및 비교
 - 코인 상세 페이지에서 소속 생태계 명확히 표시
+
+> 현재 개발 예정 기능입니다.
 
 ### 3.3 기술적 지표 및 신호
