+++ backend/src/main/java/com/crypto/market/insight/domain/market/dto/FavoriteDto.java
@@ -4,13 +4,17 @@ import com.crypto.market.insight.domain.market.model.entity.Favorite;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 public class FavoriteDto {
 
     @Schema(description = "즐겨찾기 추가 요청")
-    @Getter
+    @Data
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
     public static class Request {
         @Schema(description = "코인 ID", example = "bitcoin")
         @NotBlank(message = "coinId는 필수입니다")
@@ -18,8 +22,9 @@ public class FavoriteDto {
     }
 
     @Schema(description = "즐겨찾기 응답")
-    @Getter
+    @Data
     @Builder
+    @AllArgsConstructor
     public static class Response {
         @Schema(description = "즐겨찾기 ID")
         private Long id;
