+++ backend/src/main/java/com/crypto/market/insight/domain/market/service/IndicatorService.java
@@ -13,11 +13,15 @@ import com.crypto.market.insight.domain.strategy.indicator.IndicatorCalculator.M
 import java.math.BigDecimal;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class IndicatorService {
 
     private static final int RSI_PERIOD = 14;
