+++ backend/src/main/java/com/crypto/market/insight/domain/market/service/MarketService.java
@@ -16,10 +16,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class MarketService {
 
     private static final String DEFAULT_VS_CURRENCY = "usd";
