+++ frontend/src/features/market/components/chart/index.ts
@@ -3,3 +3,5 @@ export { default as VolumeChart } from './VolumeChart';
 export { default as ChartContainer } from './ChartContainer';
 export { default as TimeframeSelector } from './TimeframeSelector';
 export { default as IndicatorSelector } from './IndicatorSelector';
+export { default as MacdPanel } from './MacdPanel';
+export { default as RsiPanel } from './RsiPanel';
