+++ frontend/src/styles/ThemeContext.ts
@@ -1,0 +1,14 @@
+import { createContext } from 'react';
+
+type ThemeMode = 'light' | 'dark';
+
+export interface ThemeContextValue {
+  mode: ThemeMode;
+  toggleTheme: () => void;
+  setTheme: (mode: ThemeMode) => void;
+  isDark: boolean;
+}
+
+export const ThemeContext = createContext<ThemeContextValue | undefined>(
+  undefined,
+);
