+++ frontend/src/styles/index.ts
@@ -1,3 +1,4 @@
 export { lightTheme, darkTheme, type Theme } from './theme';
-export { ThemeProvider, useTheme } from './ThemeProvider';
+export { ThemeProvider } from './ThemeProvider';
+export { useTheme } from './useTheme';
 export { MARKET_COLORS, CHART_COLORS } from './marketColors';
