+++ frontend/src/styles/useTheme.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ThemeContext, type ThemeContextValue } from './ThemeContext';
+
+export function useTheme(): ThemeContextValue {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+}
